+++ src/App/pages/RecipeDetails/RecipeDetails.module.scss
@@ -30,185 +30,4 @@
     margin: 0 auto;
     padding: 20px;
   }
-
-  &__top {
-    display: grid;
-    grid-template-areas:
-      "title title title"
-      "image info info"
-      "summary summary summary";
-  }
-
-  &__title {
-    grid-area: title;
-    font-weight: 700;
-    font-size: 44px;
-    line-height: 109%;
-    color: $text-primary;
-    margin-bottom: $space-36;
-    padding-left: $space-16;
-    position: relative;
-
-    svg {
-      position: absolute;
-      left: 0;
-      top: 50%;
-      transform: translate(-50%, -50%);
-    }
-  }
-
-  &__image {
-    grid-area: image;
-    border-radius: 10px;
-    width: 448px;
-    height: 298px;
-    object-fit: cover;
-    width: 100%;
-  }
-
-  &__info {
-    grid-area: info;
-    grid-column: 2/4;
-    display: flex;
-    justify-content: start;
-    flex-wrap: wrap;
-    margin-left: 150px;
-    align-items: center;
-
-    &Item {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      flex-basis: 33.33%;
-      gap: $space-xxs;
-
-      span:last-child {
-        font-weight: 600;
-        font-size: 16px;
-        line-height: 125%;
-        color: $brand;
-      }
-    }
-  }
-
-  &__summary {
-    grid-area: summary;
-    margin-top: $space-48;
-    padding: $space-36;
-    background: rgba(181, 70, 15, 0.33);
-    border-radius: 10px;
-    font-size: 16px;
-    line-height: 125%;
-    margin-bottom: $space-48;
-
-    a {
-      color: $text-primary;
-      text-decoration: none;
-
-      &:link {
-        text-decoration: underline;
-      }
-
-      &:hover {
-        text-decoration: none;
-      }
-    }
-  }
-
-  &__necessity {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: $space-48;
-    margin-bottom: $space-48;
-
-    h2 {
-      font-weight: 600;
-      font-size: 20px;
-      line-height: 120%;
-      color: $text-primary;
-      margin-bottom: $space-xl;
-    }
-
-    & .recipe__section {
-      margin-bottom: $space-48;
-
-      ul {
-        list-style: none;
-        display: flex;
-        gap: 10px;
-        flex-wrap: wrap;
-        font-size: 16px;
-        line-height: 125%;
-
-        li {
-          flex-basis: 48.8%;
-          position: relative;
-          padding-left: 35px;
-        }
-      }
-      &-ingredients {
-
-        li:before {
-          content: "";
-          position: absolute;
-          left: 0;
-          transform: translateY(-50%, -50%);
-          width: 25px;
-          height: 17px;
-
-          background:url("@assets/dish.svg") no-repeat center center / contain;
-        }
-      }
-      &-equipment {
-        li:before {
-          content: "";
-          position: absolute;
-          left: 0;
-          top: 0;
-
-          width: 25px;
-          height: 25px;
-          background: $brand;
-          background: url("assets/object.svg") no-repeat center center / contain;
-          background: url("@assets/object.svg") no-repeat center center / contain;
-        }
-      }
-
-      &:last-child {
-        li:before {
-          
-        }
-      }
-    }
-  }
-
-  &__section-steps {
-    margin-bottom: $space-16;
-
-    h2 {
-      font-weight: 600;
-      font-size: 20px;
-      line-height: 120%;
-      color: $text-primary;
-      margin-bottom: $space-xl;
-    }
-
-    display: flex;
-    flex-direction: column;
-    gap: $space-16;
-
-    h3 {
-      font-weight: 600;
-
-      line-height: 125%;
-      color: $text-primary;
-      margin-bottom: $space-xxs;
-    }
-
-    p {
-      font-size: 14px;
-      line-height: 125%;
-      color: $text-primary;
-    }
-  }
 }
