+++ src/App/pages/RecipeDetails/components/RecipeNecessity/RecipeNecessity.module.scss
@@ -1,0 +1,85 @@
+.recipe__necessity {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: $space-48;
+  margin-bottom: $space-48;
+
+  h2 {
+    font-weight: 600;
+    font-size: 20px;
+    line-height: 120%;
+    color: $text-primary;
+    margin-bottom: $space-xl;
+  }
+
+  .recipe__section {
+    margin-bottom: $space-48;
+
+    ul {
+      list-style: none;
+      display: flex;
+      gap: 10px;
+      flex-wrap: wrap;
+      font-size: 16px;
+      line-height: 125%;
+
+      li {
+        flex-basis: 48.8%;
+        position: relative;
+        padding-left: 35px;
+      }
+    }
+
+    &-ingredients {
+      li:before {
+        content: "";
+        position: absolute;
+        left: 0;
+        transform: translateY(-50%, -50%);
+        width: 25px;
+        height: 17px;
+        background: url("@assets/dish.svg") no-repeat center center / contain;
+      }
+    }
+
+    &-equipment {
+      li:before {
+        content: "";
+        position: absolute;
+        left: 0;
+        top: 0;
+        width: 25px;
+        height: 25px;
+        background: url("@assets/object.svg") no-repeat center center / contain;
+      }
+    }
+  }
+}
+
+.recipe__section-steps {
+  margin-bottom: $space-16;
+  display: flex;
+  flex-direction: column;
+  gap: $space-16;
+
+  h2 {
+    font-weight: 600;
+    font-size: 20px;
+    line-height: 120%;
+    color: $text-primary;
+    margin-bottom: $space-xl;
+  }
+
+  h3 {
+    font-weight: 600;
+    line-height: 125%;
+    color: $text-primary;
+    margin-bottom: $space-xxs;
+  }
+
+  p {
+    font-size: 14px;
+    line-height: 125%;
+    color: $text-primary;
+  }
+}
