+++ src/App/pages/RecipeDetails/components/RecipeNecessity/index.ts
@@ -1,0 +1,3 @@
+export * from './RecipeNecessity.tsx';
+
+
