+++ src/App/pages/RecipeDetails/components/RecipeTop/RecipeTop.module.scss
@@ -1,0 +1,85 @@
+.recipe {
+  &__top {
+    display: grid;
+    grid-template-areas:
+      "title title title"
+      "image info info"
+      "summary summary summary";
+  }
+
+  &__title {
+    grid-area: title;
+    font-weight: 700;
+    font-size: 44px;
+    line-height: 109%;
+    color: $text-primary;
+    margin-bottom: $space-36;
+    padding-left: $space-16;
+    position: relative;
+
+    svg {
+      position: absolute;
+      left: 0;
+      top: 50%;
+      transform: translate(-50%, -50%);
+    }
+  }
+
+  &__image {
+    grid-area: image;
+    border-radius: 10px;
+    width: 448px;
+    height: 298px;
+    object-fit: cover;
+    width: 100%;
+  }
+
+  &__info {
+    grid-area: info;
+    grid-column: 2/4;
+    display: flex;
+    justify-content: start;
+    flex-wrap: wrap;
+    margin-left: 150px;
+    align-items: center;
+
+    &Item {
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+      flex-basis: 33.33%;
+      gap: $space-xxs;
+
+      span:last-child {
+        font-weight: 600;
+        font-size: 16px;
+        line-height: 125%;
+        color: $brand;
+      }
+    }
+  }
+
+  &__summary {
+    grid-area: summary;
+    margin-top: $space-48;
+    padding: $space-36;
+    background: rgba(181, 70, 15, 0.33);
+    border-radius: 10px;
+    font-size: 16px;
+    line-height: 125%;
+    margin-bottom: $space-48;
+
+    a {
+      color: $text-primary;
+      text-decoration: none;
+
+      &:link {
+        text-decoration: underline;
+      }
+
+      &:hover {
+        text-decoration: none;
+      }
+    }
+  }
+}
