+++ src/App/pages/RecipeDetails/components/RecipeTop/index.ts
@@ -1,0 +1,1 @@
+export * from './RecipeTop.tsx';
