+++ src/App/pages/RecipesList/components/Pagination/index.ts
@@ -1,1 +1,1 @@
-export { default } from './Pagination.tsx';
+export { default } from './Pagination';
