--- src/App/pages/components/LoaderContainer/LoaderContainer.ts
@@ -1,1 +1,0 @@
-export { default } from './LoaderContainer.tsx';
