+++ src/App/pages/components/LoaderContainer/index.ts
@@ -1,0 +1,1 @@
+export { default } from './LoaderContainer';
